Fix SharedInvalidationMessage struct.
SharedInvalidationMessage, and all the other structs that are part of
the union, must agree on the width of the shared 'id' field.

+++ src/include/storage/sinval.h
@@ -63,7 +63,7 @@
 typedef struct
 {
 	/* note: field layout chosen with an eye to alignment concerns */
-	int32		id;				/* cache ID --- must be first */
+	int16		id;				/* cache ID --- must be first */
 	ItemPointerData tuplePtr;	/* tuple identifier in cached relation */
 	Oid			dbId;			/* database ID, or 0 if a shared relation */
 	uint32		hashValue;		/* hash value of key for this catcache */
@@ -82,7 +82,7 @@ typedef struct
 
 typedef struct
 {
-	int32		id;				/* type field --- must be first */
+	int16		id;				/* type field --- must be first */
 	Oid			dbId;			/* database ID, or 0 if a shared relation */
 	Oid			relId;			/* relation ID */
 } SharedInvalRelcacheMsg;
@@ -91,7 +91,7 @@ typedef struct
 
 typedef struct
 {
-	int32		id;				/* type field --- must be first */
+	int16		id;				/* type field --- must be first */
 	RelFileNode rnode;			/* physical file ID */
 } SharedInvalSmgrMsg;
 
@@ -105,7 +105,7 @@ typedef struct
 
 typedef union
 {
-	int32		id;				/* type field --- must be first */
+	int16		id;				/* type field --- must be first */
 	SharedInvalCatcacheMsg cc;
 	SharedInvalCatalogMsg cat;
 	SharedInvalRelcacheMsg rc;
